# excel-api-node/resources/openapi.yaml
openapi: "3.1.0"
info:
  title: "Excel File API"
  description: "HTTP service exposing Excel spreadsheet data as a JSON API"
  version: "0.0.3"
  contact:
    name: "Filip Golewski"
  license:
    name: "MIT"
servers:
  - url: "http://localhost:8443/api/v1"
    description: "Local development"
security:
  - bearerJwt: []
  - staticToken: []
paths:
...
  /workbooks/{fileId}/sheets/{sheetName}:
    get:
      summary: "Sheet metadata"
      operationId: "getSheet"
      tags: ["Sheets"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
      responses:
        "200":
          description: "Sheet metadata"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetInfo"
        "404":
          description: "Workbook or sheet not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /workbooks/{fileId}/sheets/{sheetName}/cells/{cellRef}:
    get:
      summary: "Read cell"
      operationId: "getCell"
      tags: ["Cells"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - $ref: "#/components/parameters/cellRef"
        - name: format
          in: query
          schema:
            type: string
            enum: ["native", "display", "string"]
            default: "native"
      responses:
        "200":
          description: "Cell data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellData"
    put:
      summary: "Write cell"
      operationId: "putCell"
      tags: ["Cells"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - $ref: "#/components/parameters/cellRef"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CellWriteRequest"
      responses:
        "200":
          description: "Cell updated"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CellData"
  /workbooks/{fileId}/sheets/{sheetName}/ranges/{rangeRef}:
    get:
      summary: "Read range"
      operationId: "getRange"
      tags: ["Cells"]
      parameters:
        - $ref: "#/components/parameters/fileId"
        - $ref: "#/components/parameters/sheetName"
        - $ref: "#/components/parameters/rangeRef"
        - name: format
          in: query
          schema:
            type: string
            enum: ["native", "display", "string"]
            default: "native"
      responses:
        "200":
          description: "Range data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RangeData"
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
    staticToken:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: Token {value}"
    oauth2:
      type: oauth2
      flows:
        password:
          tokenUrl: /auth/token
          scopes:
            read: "Read workbook data"
            write: "Modify workbook data"
            admin: "Administrative operations"
        clientCredentials:
          tokenUrl: /auth/token
          scopes:
            read: "Read workbook data"
            write: "Modify workbook data"
            admin: "Administrative operations"
  parameters:
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
      description: "Workbook identifier from registry"
    sheetName:
      name: sheetName
      in: path
      required: true
      schema:
        type: string
      description: "Worksheet name (URL-encoded)"
...
    cellRef:
      name: cellRef
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Z]{1,3}[0-9]+$"
      description: "Cell reference in Excel notation (e.g. A1, B5, AA100)"
    rangeRef:
      name: rangeRef
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Z]{1,3}[0-9]+:[A-Z]{1,3}[0-9]+$"
      description: "Range in Excel notation (e.g. A1:D10)"
  schemas:
...
    SheetInfo:
      type: object
      required: ["name", "row_count", "column_count"]
      properties:
        name:
          type: string
        row_count:
          type: integer
        column_count:
          type: integer
        mode:
          type: string
          enum: ["table", "raw"]
        header_row:
          type: integer
        first_data_row:
          type: integer
...
    CellData:
      type: object
      required: ["ref", "value", "type"]
      properties:
        ref:
          type: string
        column:
          type: string
        row:
          type: integer
        value: {}
        type:
          type: string
          enum: ["string", "number", "date", "boolean", "formula", "empty"]
        number_format:
          type: string
          nullable: true
        is_formula:
          type: boolean
        formatted:
          type: string
          nullable: true
    CellWriteRequest:
      type: object
      required: ["value"]
      properties:
        value: {}
    RangeData:
      type: object
      required: ["range", "rows"]
      properties:
        range:
          type: string
        rows:
          type: array
          items:
            type: object
            required: ["row", "cells"]
            properties:
              row:
                type: integer
              cells:
                type: array
                items:
                  $ref: "#/components/schemas/CellData"
...
    ErrorResponse:
      type: object
      required: ["error", "message"]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true